Add observeAddingTalkArea.js
